+++ docs/content/api_core.rst
@@ -159,11 +159,11 @@ to get specific sequences from your genome, and save these to file.
 ------------
 
 You can obtain genomic sequence from a wide variety of inputs with :code:`as_seqdict`.
-To use the function, it must be specifically imported with :code:`from genomepy.genome.seqdict import as_seqdict`.
+To use the function, it must be explicitly imported with :code:`from genomepy.seq import as_seqdict`.
 
 ------------
 
-.. autofunction:: genomepy.genome.seqdict.as_seqdict
+.. autofunction:: genomepy.seq.as_seqdict
    :noindex:
 
 ------------
